backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20

# RAG & LLM
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-text-splitters==0.3.6
agno==1.0.0
ollama==0.4.7

# Vector Store & Embeddings
qdrant-client==1.13.2
langchain-qdrant==0.2.0
fastembed==0.5.1

# OCR & Document Processing
Pillow==10.4.0
python-docx==0.8.11
PyPDF2==3.0.1
PyMuPDF==1.23.5
pytesseract==0.3.10

# Dots.OCR Dependencies
transformers>=4.40.0
torch>=2.0.0
qwen-vl-utils>=0.0.11
accelerate>=0.20.0
packaging>=21.0
# flash-attn>=2.0.0  # Optional - can be installed separately if needed

# Hebrew Language Processing
hebrew-tokenizer==2.3.0

# Audio Processing
librosa

# System Monitoring
psutil==5.9.6

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.28.1
numpy==1.26.4
pandas==2.1.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
isort==5.12.0